% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_agenda_item.R
\name{browse_agenda_item}
\alias{browse_agenda_item}
\title{browse aganda item}
\usage{
browse_agenda_item(nativeTermYear, referenceNumber)
}
\arguments{
\item{nativeTermYear}{int}

\item{referenceNumber}{char}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
opens an agneda item in a browser
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 browse_agenda_item(2022, "BA198.1")
 }
}
}
\seealso{
\code{\link[utils]{browseURL}}
}
