% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_terms.R
\name{car_terms}
\alias{car_terms}
\title{car terms}
\usage{
car_terms()
}
\value{
tibble with two columns: sector and key_terms
}
\description{
returns a lookup table for car-related terms
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 car_terms()
 }
}
}
\seealso{
\code{\link[snakecase]{caseconverter}}
 \code{\link[tibble]{tribble}}
 \code{\link[tidyr]{unnest_longer}}
 \code{\link[dplyr]{arrange}}
}
