% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_agendas.R
\name{categorize_agendas}
\alias{categorize_agendas}
\title{categorize agenda}
\usage{
categorize_agendas(decisionBodyId)
}
\arguments{
\item{decisionBodyId}{integer}
}
\value{
list
}
\description{
categorizes all agenda items for a decisionBodyId
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 results <- categorize_agendas(2467)
 results <- categorize_agendas(2566)
 View(results$categorized)
 View(results$unmatched)

 }
}
}
\seealso{
\code{\link[purrr]{map}}, \code{\link[purrr]{list_c}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{filter}}
 \code{\link[stringr]{str_detect}}, \code{\link[stringr]{modifiers}}
 \code{\link[tibble]{tibble}}
 \code{\link[tidyr]{unnest}}
}
