% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize_df}
\alias{categorize_df}
\title{categorize df}
\usage{
categorize_df(df, name_col = "agendaItemTitle", terms_tbl = NULL)
}
\arguments{
\item{df}{dataframe}

\item{name_col}{char, Default: 'agendaItemTitle'}

\item{terms_tbl}{LUT as a tibble, Default: key_terms()}
}
\value{
tibble
}
\description{
catagorize a dataframe using a named column
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 example_data <- tibble::tribble(
 ~agendaItemTitle, ~id,
"Greening the military must align with Climate action", 1,
"Deadly nordic heatwave supercharged by climate crisis", 2,
"Southwest Airlines sells renewable unit in retreat form climate goals", 3)
categorize_df(example_data)
 }
}
}
\seealso{
\code{\link[torontodata:categorize_txt]{categorize_txt}},
}
