% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_text.R
\name{categorize_text}
\alias{categorize_text}
\title{catgeorize text}
\usage{
categorize_text(text)
}
\arguments{
\item{text}{text}
}
\value{
tibble
}
\description{
catagorizes a text using key_terms()
}
\details{
DETAILS
}
\examples{
categorize_text("Climate change is real")
}
\seealso{
\code{\link[stringr]{str_detect}}, \code{\link[stringr]{modifiers}}
 \code{\link[tibble]{tibble}}
}
