% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize_txt}
\alias{categorize_txt}
\title{categorize text}
\usage{
categorize_txt(text, terms_tbl = key_terms())
}
\arguments{
\item{text}{char, a character vector}

\item{terms_tbl}{a lut as a tibble, Default: key_terms()}
}
\value{
tibble
}
\description{
using a look up table (LUT), such as one generated by
car_terms() or key_terms(), return a datframe with the text, key_term,
pattern, sector, match and keyword_count
}
\details{
DETAILS
}
\examples{
 txt = "As Green wild fires choke UP with conservation, pedestrians demand
 cooling centres. But not climate change."
 categorize_txt(txt)
}
\seealso{
\code{\link[torontodata:categorize_df]{categorize_df}},
}
