% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut_climate_terms.R
\name{climate_terms}
\alias{climate_terms}
\title{key terms}
\usage{
climate_terms()
}
\value{
tibble
}
\description{
creates a tibble to use as a look u table
}
\details{
based on climatefast's "Key terms list-2.pdf"
}
\examples{
\dontrun{
if(interactive()){
 View(climate_terms())
 }
}
}
\seealso{
\code{\link[tibble]{tribble}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}
 \code{\link[purrr]{map}}
 \code{\link[tidyr]{unnest_longer}}
}
