% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/council_date.R
\name{council_date}
\alias{council_date}
\title{council date}
\usage{
council_date(nativeTermYear = 2025, referenceNumber = "IE23.2")
}
\arguments{
\item{nativeTermYear}{integer}

\item{referenceNumber}{chraacter}
}
\value{
charcater vector
}
\description{
returns the date that follows the string
 *It will be considered by City Council on
 in agend item webspages with a URL of the form
 https://secure.toronto.ca/council/agenda-item.do?item=nativeTermYear,referenceNumber
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 # council_date(nativeTermYear = 2025,  referenceNumber = "IE23.2")
 }
}
}
\seealso{
\code{\link[rvest]{read_html}}, \code{\link[rvest]{html_element}}, \code{\link[rvest]{html_text}}
 \code{\link[stringr]{str_detect}}
}
