% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_open_dataR.R
\name{download_voting_records}
\alias{download_voting_records}
\title{download voting records}
\usage{
download_voting_records(dest_dir = "inst/extdata/to_open_data")
}
\arguments{
\item{dest_dir}{path, Default: 'inst/extdata/to_open_data'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
downloadsmember_vorting_records from toronto open data
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
