% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_all_agenda_items.R
\name{fetch_agenda_item}
\alias{fetch_agenda_item}
\title{fetch agenda item}
\usage{
fetch_agenda_item(meetingId)
}
\arguments{
\item{meetingId}{integer meetingId}
}
\value{
data frame
}
\description{
from meeting, get the dataframe $sections$agendeItems[[1]]
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{pluck}}
}
