% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_agenda_item_council_date.R
\name{fetch_agenda_item_council_date}
\alias{fetch_agenda_item_council_date}
\title{fetch agenda council date}
\usage{
fetch_agenda_item_council_date(
  url = "https://secure.toronto.ca/council/agenda-item.do?item=2025.EX24.2"
)
}
\arguments{
\item{url}{a valid url,
Default: 'https://secure.toronto.ca/council/agenda-item.do?item=2023.IE1.1'}
}
\value{
date
}
\description{
scrapes a website to fetch that date that an item will be
 discussed in Council
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 fetch_agenda_item_council_date(
  "https://secure.toronto.ca/council/agenda-item.do?item=2025.EX24.2"
   )
 }
}
}
\seealso{
\code{\link[rvest]{read_html}},
 \code{\link[rvest]{html_element}},
 \code{\link[rvest]{html_text}}
 \code{\link[stringr]{str_detect}}
}
