% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_agenda_item_status.R
\name{fetch_agenda_item_status}
\alias{fetch_agenda_item_status}
\title{fetch_agenda_item_status}
\usage{
fetch_agenda_item_status(meetingId = 25775)
}
\arguments{
\item{meetingId}{integer}
}
\value{
dataframe
}
\description{
FUNCTION_DESCRIPTION
}
\details{
fetches the date that an item will be discussed in a council meeting
 with fetch_individual_meeting() and updates the dataframe with the result of
 a query that attempts to get the council data by scraping the website for
 the agendaItem.
}
\examples{
\dontrun{
if(interactive()){
 fetch_agenda_item_status(meetingId = 25775)
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
