% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_all_agenda_items.R
\name{fetch_all_agenda_items}
\alias{fetch_all_agenda_items}
\title{fetch all agenda_items}
\usage{
fetch_all_agenda_items(decisionBodyId)
}
\arguments{
\item{decisionBodyId}{integer, decisonBodyId}
}
\value{
data frame
}
\description{
fetch all agendaItems for a decisionBody
}
\details{
gets all the meeitngIds for a decsionBody,
then all the agendaItems,
converts the condiderStartTime from millisonds into a POSIX date
}
\examples{
\dontrun{
if(interactive()){
 agenda_items_29784 <- fetch_all_agenda_items(decisionBodyId = 2984)
 }
}
}
\seealso{
\code{\link[dplyr]{pull}}, \code{\link[dplyr]{mutate}}
 \code{\link[purrr]{map}}, \code{\link[purrr]{keep}}, \code{\link[purrr]{list_c}}
 \code{\link[lubridate]{as_date}}
}
