% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_individual_meeting_agenda_item_titles.R
\name{fetch_individual_meeting_agenda_item_titles}
\alias{fetch_individual_meeting_agenda_item_titles}
\title{Fetch agenda item titles from a meeting}
\usage{
fetch_individual_meeting_agenda_item_titles(meetingId)
}
\arguments{
\item{meetingId}{Numeric ID of the meeting}
}
\value{
a dataframe with nativeTermYear, referenceNumber, agendaItemTitle
}
\description{
Fetch agenda item titles from a meeting
}
\examples{
# example code
fetch_individual_meeting_agenda_item_titles(25793)
fetch_individual_meeting_agenda_item_titles(meetingId = 25775)

}
