% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_multiple_decisionbody_list.R
\name{fetch_multiple_decisionbody_list}
\alias{fetch_multiple_decisionbody_list}
\title{fetch a list of decision bodies from the Toronto Council API}
\usage{
fetch_multiple_decisionbody_list(term_id = 8)
}
\arguments{
\item{term_id}{the current term as an integer,}
}
\value{
a data frame of decison bodies
}
\description{
This function retrieves a list of decision-making bodies for a given term, i.e. (8 is 2022-2026)
}
\examples{
fetch_multiple_decisionbody_list()
}
