% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_multiple_meeting.R
\name{fetch_multiple_meeting}
\alias{fetch_multiple_meeting}
\title{fetch_multiple_meeting}
\usage{
fetch_multiple_meeting(decisionBodyId = 2566)
}
\arguments{
\item{decisionBodyId}{integer, Default: 2566}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 meetings_2566 <- fetch_multiple_meeting(decisionBodyId = 2566)
 }
}
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
}
