% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_terms.R
\name{key_terms}
\alias{key_terms}
\title{key terms}
\usage{
key_terms()
}
\value{
tibble
}
\description{
creates a tibble
}
\details{
based on "Key terms list-2.pdf"
}
\examples{
\dontrun{
if(interactive()){
 View(key_terms())
 }
}
}
\seealso{
\code{\link[tibble]{tribble}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{arrange}}
 \code{\link[purrr]{map}}
 \code{\link[tidyr]{unnest_longer}}
}
