% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_decision_body_id.R
\name{lookup_decision_body_id}
\alias{lookup_decision_body_id}
\title{lookup decision body}
\usage{
lookup_decision_body_id(name = "City Council")
}
\arguments{
\item{name}{PARAM_DESCRIPTION, Default: 'City Council'}
}
\value{
integer
}
\description{
look up the decisionBodyId of a Committee (including City Council) by name
}
\details{
looks up the decisionBodyId in fetch_multiple_decisionbody_list()
}
\examples{
\dontrun{
if(interactive()){
 lookup_decision_body_id("Executive Committee")
 }
}
}
\seealso{
\code{\link[torontodata:fetch_multiple_decisionbody_list]{fetch_multiple_decisionbody_list}}
}
