% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut_decisionbody_id_code_name.R
\name{lut_decisionbody_id_code_name}
\alias{lut_decisionbody_id_code_name}
\title{data decision body title}
\usage{
lut_decisionbody_id_code_name()
}
\value{
tibble
}
\description{
returns a tibble with the decisionBodyId, decisionBodyCode and decisionBodyName for the Toronto Council commitees
}
\details{
decisionBodyCode is not something that exists in the API, but
something I made up that is the first two (or three!, careful) LETTERS
of the referenceNumber of an agenda item
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tribble}}
}
