% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_individual_meeting_agenda_item_titles.R
\name{make_agenda_item_title_fetcher}
\alias{make_agenda_item_title_fetcher}
\title{make_agenda_item_title_fetcher}
\usage{
make_agenda_item_title_fetcher()
}
\value{
fiction
}
\description{
returns a closure that caches
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
  agenda_fetcher <- make_agenda_item_title_fetcher()
  agenda_fetcher(24392)
  agenda_fetcher(23179)
  agenda_fetcher(24392)
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
