% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_member_voting_record.R
\name{make_member_voting_record}
\alias{make_member_voting_record}
\title{make member voting record}
\usage{
make_member_voting_record()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
creates a closure
}
\examples{
\dontrun{
if(interactive()){
 get_member_voting_record <- make_member_voting_record()
get_member_voting_record(2023, "BU3.1")
 }
}
}
