% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_counter.R
\name{make_safe_vote_counter}
\alias{make_safe_vote_counter}
\title{Fmake safe vote counter}
\usage{
make_safe_vote_counter()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
closure-based caching function to count votes in Toronto Council agenda items,
expects a h4 heading with class <h4 class="vote-table-header"> followed  by a table
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 vote_counter <- make_safe_vote_counter()
vote_counter(2022, "BA198.1", use_cache = FALSE)
vote_counter(2011, "EX10.1")
vote_counter(2011, "EX10.1")
 }
}
}
