% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_counter.R
\name{make_vote_counter}
\alias{make_vote_counter}
\title{make_vote_counter}
\usage{
make_vote_counter()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
closure-based caching version of vote counter
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 vote_counter <- make_vote_counter()
 vote_counter(2023, "MM6.17")
 }
}
}
