% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_key_terms.R
\name{match_key_terms}
\alias{match_key_terms}
\title{match key terms}
\usage{
match_key_terms(df, name_col = "agendaItemTitle", lut = key_terms())
}
\arguments{
\item{df}{data frame}

\item{name_col}{character, Default: 'agendaItemTitle'}

\item{lut}{data frame, Default: key_terms()}
}
\value{
data frame
}
\description{
use a lookup table to match key tersma snf sectors to agendaItemTitle
}
\details{
this function assumes a dataframe that contains a title or text
string that contains climate-related issues
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
