% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{match_key_terms_core}
\alias{match_key_terms_core}
\title{match key terms}
\usage{
match_key_terms_core(text_vec, terms_tbl = key_terms())
}
\arguments{
\item{text_vec}{char}

\item{terms_tbl}{tibble, Default: key_terms()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
utility function for categorization using a LUT
}
\details{
this gets called from categorize_txt() or categorize_df()
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[torontodata:categorize_df]{categorize_txt}}
 \code{\link[torontodata:categorize_df]{categorize_df}}
}
