% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_key_terms.R
\name{match_key_terms_tidy}
\alias{match_key_terms_tidy}
\title{match_key_terms - tidy}
\usage{
match_key_terms_tidy(df, name_col = "agendaItemTitle", lut = key_terms())
}
\arguments{
\item{df}{daatrframe}

\item{name_col}{name of a column that contains words to be macthed, Default: 'agendaItemTitle'}

\item{lut}{a look up tableas a tibble of dataframe PARAM_DESCRIPTION, Default: key_terms()}
}
\value{
data frame
}
\description{
updatye a dtaframe with matched key_terms
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
df <- tibble::tribble( ~agendaItemTitle,  "Trees are cool", "Streets are hot")
match_key_terms_tidy(df)
 }
}
}
