% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torontodata-package.R
\docType{package}
\name{torontodata}
\alias{torontodata-package}
\alias{torontodata}
\title{torontodata: Tools for accessing and working with Toronto Council data}
\description{
This package provides functions and datasets for interacting with the Toronto Council API,
including decision body lookups, term-based queries, and cached API access.
}
\author{
\strong{Maintainer}: Michiel Duvekot \email{michiel@duvekot.ca}

}
