% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_agendas.R
\name{categorize_agendas}
\alias{categorize_agendas}
\title{categorize agenda}
\usage{
categorize_agendas(decisionBodyId)
}
\arguments{
\item{decisionBodyId}{integer}
}
\value{
list
}
\description{
categorizes all agenda items for a decisionBodyId
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 results <- categorize_agendas(2468)
 results <- categorize_agendas(2566)
 results <- categorize_agendas(2945)
 results <- categorize_agendas(2946)

decisionBodyId = 2945
fetch_individual_meeting_agenda_item_titles(2945)
meetingId = 26585
meetingId = 27110
 View(results$categorized)
 View(results$unmatched)

 }
}
}
