% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_text.R
\name{categorize_text}
\alias{categorize_text}
\title{categorize text}
\usage{
categorize_text(text, terms_tbl = key_terms())
}
\arguments{
\item{text}{text}

\item{terms_tbl}{tibble with two columnsl key_terms and sector. Default: key_terms()}
}
\value{
tibble
}
\description{
categorizes a text using key_terms()
}
\details{
DETAILS
}
\examples{
categorize_text(text = "The Toronto Atmospheric Fund's (TAF) mandate
includes the promotion of global climate stabilization through the
reduction of greenhouse gas emissions, local air quality, energy
conservation and efficiency, public understanding of global warming
and its implications for the urban environment, and related research
and technology development. ")
}
\seealso{
\code{\link[stringr]{str_detect}}, \code{\link[stringr]{modifiers}}
 \code{\link[tibble]{tibble}}
}
