% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_councillors.R
\name{fetch_councillors}
\alias{fetch_councillors}
\title{FUNCTION_TITLE}
\usage{
fetch_councillors()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 get_councillors <- fetch_councillors()
get_councillors()|> print (n = Inf)
get_councillors()$Councillor
get_councillors()$last_name
 }
}
}
\seealso{
\code{\link[rvest]{read_html}}, \code{\link[rvest]{html_table}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{c("rowwise", "rowwise")}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{arrange}}
 \code{\link[stringr]{str_replace}}, \code{\link[stringr]{str_split}}
}
