% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_individual_meeting.R
\name{fetch_individual_meeting}
\alias{fetch_individual_meeting}
\title{fetch individual meeting}
\usage{
fetch_individual_meeting(meetingId = 25849)
}
\arguments{
\item{meetingId}{integer, Default: 25849}
}
\value{
OUTPUT_DESCRIPTION

list
}
\description{
FUNCTION_DESCRIPTION

fetches the Record from URLS like "https://secure.toronto.ca/council/api/individual/meeting/25849.json"
}
\details{
DETAILS

fetch JSON data from the toronto council API for a given meeetingId,
  as you might see in URLs like https://secure.toronto.ca/council/#/committees/2468/25777
  where 2468 is the decisionBodyId and 25777 the meetingId
  returns the Record, which contains a list of 5 items,
  the one you want is most likely the sections, which contains agendaItems
}
\examples{
\dontrun{
if(interactive()){
 fetch_individual_meeting(meetingId = 25849)
 }
}
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
}
