% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_individual_meeting.R
\name{fetch_individual_meeting}
\alias{fetch_individual_meeting}
\title{fetch_individual_meeting}
\usage{
fetch_individual_meeting(meetingId = 25849)
}
\arguments{
\item{meetingId}{integer, Default: 25849}
}
\value{
list
}
\description{
fetches the Record from URLS like "https://secure.toronto.ca/council/api/individual/meeting/25849.json"
}
\details{
DETAILS
}
\examples{
meeting_data <- fetch_individual_meeting(meetingId = 25849)
agendaItemTitles <- meeting_data$sections$agendaItems |>
  purrr::map(~ dplyr::pull(.x, agendaItemTitle)) |>
  purrr::flatten_chr()
agendaItemTitles
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}
}
