% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_decision_body_id.R
\name{lookup_decision_body_id}
\alias{lookup_decision_body_id}
\title{lookup the decisionbodyId  of a city Committee (including Council) by name}
\usage{
lookup_decision_body_id(name = "City Council")
}
\arguments{
\item{name, }{the exact name of the decisionBody}
}
\value{
an inter value for the decisionBodyId
}
\description{
This function retrieves a an interger value for a decisionBody
}
\examples{
lookup_decision_body_id()
}
