% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_key_terms.R
\name{match_key_terms}
\alias{match_key_terms}
\title{FUNCTION_TITLE}
\usage{
match_key_terms(df, name_col = "agendaItemTitle", lut = key_terms)
}
\arguments{
\item{df}{dataframe}

\item{name_col}{character, the anme of the column that tas the text t0 be matched}

\item{lut}{a dataframe, like key_terms()}
}
\value{
OUTPUT_DESCRIPTIONdtaframe
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 df <- fetch_all_agenda_items(decisionBodyId = 2984)
match_key_terms(df, "agendaItemTitle", key_terms())
 }
}
}
